requirements.txt
discord.py==1.5.0
requests==2.25.1
Pillow==8.0.1
pytz==2020.5
